restored google_snippets
